rename cli pkg
