+++ cmd/bankid/cli/prompts.go
@@ -1,4 +1,4 @@
-package src
+package cli
 
 import (
 	"path/filepath"
