+++ cmd/bankid/cli/terminal.go
@@ -1,4 +1,4 @@
-package src
+package cli
 
 import (
 	"context"
