+++ cmd/bankid/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/nicolaa5/bankid/cmd/bankid/cli"
+)
+
+func main() {
+	cli.Execute()
+}
