--- cmd/cli/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/nicolaa5/bankid/cmd/cli/src"
-)
-
-func main() {
-	src.Execute()
-}
